Zero out ISO keyword sequence for BANT0086
Although the SD2 .Install file for BANT0086 (Bosch immobilizer module) gives an ordinary-looking ISO keyword sequence, this ECU apparently does not do slow-init (or any other init) so there is no opportunity for the ISO keyword bytes to be read during this phase. The .ecu file on the SD2 simply sends six zeroes for the ISO sequence, so we want to replicate that here.